not found error problem here

+++ src/app/Modules/Specialities/Specialities.service.ts
@@ -1,5 +1,7 @@
 import { UserStatus } from "@prisma/client";
 import prisma from "../../shared/prisma";
+import ApiError from "../../Error-Handler/ApiError";
+import { StatusCodes } from "http-status-codes";
 
 const createSpecialtiesDB = async (tokenId: string, payload: any) => {
   await prisma.user.findUniqueOrThrow({
@@ -35,6 +37,20 @@ const deleteSpecialtiesDB = async (tokenId: string, specialtiesId: string) => {
       isDelete: false,
     },
   });
+
+
+  //! there is a problem here not found error no provide
+  const specialties = await prisma.specialties.findUniqueOrThrow({
+    where: {
+      id: specialtiesId,
+    },
+  });
+  console.log(specialties);
+  
+
+  if (!specialties) {
+    throw new ApiError(StatusCodes.NOT_FOUND, "No found")
+  }
 
   const result = await prisma.specialties.delete({
     where: {
